fix: обработать ошибки в запросах

+++ controllers/cards.js
@@ -1,5 +1,6 @@
 const card = require('../models/card');
-const trowError = require('../utils/errors');
+const sendError = require('../utils/errors');
+const ObjectId = require('mongoose').Types.ObjectId;
 
 module.exports.createCard = (req, res) => {
 
@@ -19,43 +20,72 @@ module.exports.createCard = (req, res) => {
     createAt
   })
     .then(card => res.send({ data: card }))
-    .catch(err => trowError(res,err));
+    .catch(err => sendError(res, err));
 };
 
 module.exports.getCards = (req, res) => {
   card.find({})
     .populate('owner')
     .then(cards => res.send({ data: cards }))
-    .catch(err => trowError(res,err));
+    .catch(err => sendError(res, err));
 };
 
 module.exports.deleteCard = (req, res) => {
-  card.findByIdAndRemove(req.params.cardId)
-    .then(cards => res.send({ data: cards }))
-    .catch(err => trowError(res,err));
+  if (ObjectId.isValid(req.params.cardId)) {
+    card.findByIdAndRemove(req.params.cardId)
+      .then(cards => {
+        if (cards) {
+          res.send({ data: cards })
+        } else {
+          return Promise.reject({ name: 'CastError' })
+        }
+      })
+      .catch(err => sendError(res, err));
+  } else {
+    sendError(res, { name: 'ValidationError' })
+  }
 };
 
 module.exports.addLike = (req, res) => {
-  card.findByIdAndUpdate(req.params.cardId,
-    { $addToSet: { likes: req.user._id } },
-    {
-      new: true,
-      runValidators: true,
-    })
-    .then(cards => res.send({ data: cards }),)
-    .catch(err => trowError(res,err));
+  if (ObjectId.isValid(req.params.cardId)) {
+    card.findByIdAndUpdate(req.params.cardId,
+      { $addToSet: { likes: req.user._id } },
+      {
+        new: true,
+        runValidators: true,
+      })
+      .then(cards => {
+        if (cards) {
+          res.send({ data: cards })
+        } else {
+          return Promise.reject({ name: 'CastError' })
+        }
+      })
+      .catch(err => sendError(res, err));
+  } else {
+    sendError(res, { name: 'ValidationError' })
+  }
 };
 
 module.exports.removeLike = (req, res) => {
-
-  card.findByIdAndUpdate(
-    req.params.cardId,
-    { $pull: { likes: req.user._id } },
-    {
-      new: true,
-      runValidators: true,
-    }
-  )
-    .then(cards => res.send({ data: cards }))
-    .catch(err => trowError(res,err));
+  if (ObjectId.isValid(req.params.cardId)) {
+    card.findByIdAndUpdate(
+      req.params.cardId,
+      { $pull: { likes: req.user._id } },
+      {
+        new: true,
+        runValidators: true,
+      }
+    )
+      .then(cards => {
+        if (cards) {
+          res.send({ data: cards })
+        } else {
+          return Promise.reject({ name: 'CastError' })
+        }
+      })
+      .catch(err => sendError(res, err));
+  } else {
+    sendError(res, { name: 'ValidationError' })
+  }
 };
